feat(var-conv-next): Add input prompt feature for naming style suggestions during variable creation
