+++ package.json
@@ -28,7 +28,8 @@
     "vscode": "^1.90.0"
   },
   "activationEvents": [
-    "*"
+    "onLanguage:javascript",
+    "onLanguage:typescript"
   ],
   "contributes": {
     "commands": [
